requirements.txt
huggingface_hub
python-dotenv
langchain 
chromadb 
sentence-transformers
langchain-community
pypdf
pymongo 
python-dotenv 

requests 
beautifulsoup4 
firecrawl
youtube-transcript-api
apscheduler
google-api-python-client
torch
pymongo
pdfminer.six
psutil
langchain-huggingface
pymupdf




